feat: error handling

+++ DummyApi/Controllers/Categories.cs
@@ -24,7 +24,15 @@ public class CategoriesController : ControllerBase
     [HttpGet("{id}")]
     public IActionResult Get(int id)
     {
-        return Ok(_dbContext.Categories.FirstOrDefault(x => x.Id == id));
+        var category = _dbContext.Categories.FirstOrDefault(x => x.Id == id);
+        if (category == null)
+        {
+            return NotFound();
+        }
+        else
+        {
+            return Ok(category);
+        }
     }
 
     // POST: /api/categories
@@ -41,10 +49,17 @@ public class CategoriesController : ControllerBase
     public IActionResult Put(int id, [FromBody] Category categoryObj)
     {
         var category = _dbContext.Categories.Find(id);
-        category.Name = categoryObj.Name;
-        category.ImageUrl = categoryObj.ImageUrl;
-        _dbContext.SaveChanges();
-        return Ok("Record updated successfully");
+        if (category == null)
+        {
+            return NotFound("No record found against this Id: " + id);
+        }
+        else
+        {
+            category.Name = categoryObj.Name;
+            category.ImageUrl = categoryObj.ImageUrl;
+            _dbContext.SaveChanges();
+            return Ok("Record updated successfully");
+        }
     }
 
     // DELETE: /api/categories/:id
@@ -52,9 +67,16 @@ public class CategoriesController : ControllerBase
     public IActionResult Delete(int id)
     {
         var category = _dbContext.Categories.Find(id);
-        _dbContext.Categories.Remove(category);
-        _dbContext.SaveChanges();
-        return Ok("Record deleted successfully");
+        if (category == null)
+        {
+            return NotFound("No record found against this Id: " + id);
+        }
+        else
+        {
+            _dbContext.Categories.Remove(category);
+            _dbContext.SaveChanges();
+            return Ok("Record deleted successfully");
+        }
     }
 
 }
